requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For Parquet file support

# HTTP requests and API clients
requests>=2.31.0
aiohttp>=3.8.0  # For async API calls

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
prophet>=1.1.0
optuna>=3.4.0  # For hyperparameter optimization
joblib>=1.3.0  # For model serialization

# Data validation and quality
pydantic>=2.0.0
marshmallow>=3.20.0

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and monitoring
loguru>=0.7.0

# Date/time handling
python-dateutil>=2.8.0

# Task scheduling and automation
schedule>=1.2.0
psutil>=5.9.0  # For process monitoring

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0